Move image implementation to separate modules.
