+++ core/tmdb-api/implementation/src/iosMain/kotlin/com/thomaskioko/tvmaniac/tmdb/implementation/TmdbPlatformComponent.kt
@@ -1,0 +1,42 @@
+package com.thomaskioko.tvmaniac.tmdb.implementation
+
+import com.thomaskioko.tvmaniac.tmdb.api.TmdbNetworkDataSource
+import com.thomaskioko.tvmaniac.util.model.Configs
+import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
+import io.ktor.client.engine.darwin.Darwin
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.json.Json
+import me.tatarka.inject.annotations.Provides
+
+actual interface TmdbPlatformComponent {
+
+    @OptIn(ExperimentalSerializationApi::class)
+    @ApplicationScope
+    @Provides
+    fun provideTmdbJson(): TmdbJson = Json {
+        isLenient = true
+        ignoreUnknownKeys = true
+        useAlternativeNames = false
+        explicitNulls = false
+    }
+
+    @ApplicationScope
+    @Provides
+    fun provideTmdbHttpClientEngine(): TmdbHttpClientEngine = Darwin.create()
+
+    @ApplicationScope
+    @Provides
+    fun provideTmdbHttpClient(
+        configs: Configs,
+        json: TmdbJson,
+        httpClientEngine: TmdbHttpClientEngine,
+    ): TmdbHttpClient = tmdbHttpClient(
+        isDebug = false,
+        json = json,
+        httpClientEngine = httpClientEngine,
+        tmdbApiKey = configs.tmdbApiKey,
+    )
+
+    @Provides
+    fun provideTmdbService(bind: TmdbNetworkDataSourceImpl): TmdbNetworkDataSource = bind
+}
