+++ data/episodeimages/api/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    id("tvmaniac.kmm.library")
+}
+
+kotlin {
+    android()
+    ios()
+
+    sourceSets {
+        sourceSets["commonMain"].dependencies {
+            api(projects.core.util)
+            api(projects.core.networkutil)
+            api(projects.core.database)
+
+            api(libs.coroutines.core)
+        }
+    }
+}
+
+android {
+    namespace = "com.thomaskioko.tvmaniac.episodeimages.api"
+}
