+++ data/episodeimages/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodeimages/api/EpisodeImageRepository.kt
@@ -1,0 +1,10 @@
+package com.thomaskioko.tvmaniac.episodeimages.api
+
+import com.thomaskioko.tvmaniac.core.networkutil.Either
+import com.thomaskioko.tvmaniac.core.networkutil.Failure
+import kotlinx.coroutines.flow.Flow
+
+interface EpisodeImageRepository {
+
+    fun updateEpisodeImage(): Flow<Either<Failure, Unit>>
+}
