+++ data/episodeimages/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodeimages/implementation/EpisodeImageComponent.kt
@@ -1,0 +1,16 @@
+package com.thomaskioko.tvmaniac.episodeimages.implementation
+
+import com.thomaskioko.tvmaniac.episodeimages.api.EpisodeImageDao
+import com.thomaskioko.tvmaniac.episodeimages.api.EpisodeImageRepository
+import me.tatarka.inject.annotations.Provides
+
+interface EpisodeImageComponent {
+
+    @Provides
+    fun provideEpisodeImageDao(bind: EpisodeImageDaoImpl): EpisodeImageDao = bind
+
+    @Provides
+    fun provideEpisodeImageRepository(
+        bind: EpisodeImageRepositoryImpl
+    ): EpisodeImageRepository = bind
+}
