+++ data/episodeimages/testing/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    id("tvmaniac.kmm.library")
+}
+
+kotlin {
+    android()
+    ios()
+
+    sourceSets {
+        sourceSets["commonMain"].dependencies {
+            implementation(projects.core.database)
+            implementation(projects.core.util)
+            implementation(projects.data.episodeimages.api)
+
+            implementation(libs.coroutines.core)
+        }
+    }
+}
+
+android {
+    namespace = "com.thomaskioko.tvmaniac.episodeimages.testing"
+}
