+++ data/episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/api/EpisodeRepository.kt
@@ -1,10 +1,3 @@
 package com.thomaskioko.tvmaniac.episodes.api
 
-import com.thomaskioko.tvmaniac.core.networkutil.Either
-import com.thomaskioko.tvmaniac.core.networkutil.Failure
-import kotlinx.coroutines.flow.Flow
-
-interface EpisodeRepository {
-
-    fun updateEpisodeArtWork(): Flow<Either<Failure, Unit>>
-}
+interface EpisodeRepository
