+++ data/episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/api/EpisodesDao.kt
@@ -1,7 +1,5 @@
 package com.thomaskioko.tvmaniac.episodes.api
 
-import com.thomaskioko.tvmaniac.core.db.EpisodeArt
-import kotlinx.coroutines.flow.Flow
 import com.thomaskioko.tvmaniac.core.db.Episodes as EpisodeCache
 
 interface EpisodesDao {
@@ -9,8 +7,6 @@ interface EpisodesDao {
     fun insert(entity: EpisodeCache)
 
     fun insert(list: List<EpisodeCache>)
-
-    fun observeEpisodeArtByShowId(): Flow<List<EpisodeArt>>
 
     fun delete(id: Long)
