--- data/episodes/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/implementation/EpisodeImageDaoImpl.kt
@@ -1,28 +1,0 @@
-package com.thomaskioko.tvmaniac.episodes.implementation
-
-import com.thomaskioko.tvmaniac.core.db.Episode_image
-import com.thomaskioko.tvmaniac.core.db.TvManiacDatabase
-import com.thomaskioko.tvmaniac.episodes.api.EpisodeImageDao
-import me.tatarka.inject.annotations.Inject
-
-@Inject
-class EpisodeImageDaoImpl(
-    private val database: TvManiacDatabase,
-) : EpisodeImageDao {
-
-    private val episodeQueries get() = database.episode_imageQueries
-
-    override fun insert(entity: Episode_image) {
-        database.transaction {
-            episodeQueries.insertOrReplace(
-                trakt_id = entity.trakt_id,
-                tmdb_id = entity.tmdb_id,
-                image_url = entity.image_url,
-            )
-        }
-    }
-
-    override fun insert(list: List<Episode_image>) {
-        list.map { insert(it) }
-    }
-}
