+++ data/episodes/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/implementation/EpisodesDaoImpl.kt
@@ -1,19 +1,13 @@
 package com.thomaskioko.tvmaniac.episodes.implementation
 
-import app.cash.sqldelight.coroutines.asFlow
-import app.cash.sqldelight.coroutines.mapToList
-import com.thomaskioko.tvmaniac.core.db.EpisodeArt
+import com.thomaskioko.tvmaniac.core.db.Episodes as EpisodeCache
 import com.thomaskioko.tvmaniac.core.db.TvManiacDatabase
 import com.thomaskioko.tvmaniac.episodes.api.EpisodesDao
-import com.thomaskioko.tvmaniac.util.model.AppCoroutineDispatchers
-import kotlinx.coroutines.flow.Flow
 import me.tatarka.inject.annotations.Inject
-import com.thomaskioko.tvmaniac.core.db.Episodes as EpisodeCache
 
 @Inject
 class EpisodesDaoImpl(
     private val database: TvManiacDatabase,
-    private val dispatchers: AppCoroutineDispatchers,
 ) : EpisodesDao {
 
     private val episodeQueries get() = database.episodesQueries
@@ -37,11 +31,6 @@ class EpisodesDaoImpl(
     override fun insert(list: List<EpisodeCache>) {
         list.map { insert(it) }
     }
-
-    override fun observeEpisodeArtByShowId(): Flow<List<EpisodeArt>> =
-        episodeQueries.episodeArt()
-            .asFlow()
-            .mapToList(dispatchers.io)
 
     override fun delete(id: Long) {
         episodeQueries.delete(id)
